Added Bitwise operator parsing
Changed the grammar files to match the correct naming of the operators.

Changed the order of the operators in TokenInfo.java so that it is easier to read.

Changed the ordering of the operators test so that unary operators are together.

Changed the Scanner token names to better reflect what is in the grammar file.

Added level 9 -> level 7 parsing and changed the level 10 and level 6 entry/exit points so that it flows correctly.

Added a BitwiseExpression which expects two variables and an operand (for example 1 | 3)